removed content from app.jsx
